requirements.txt
Flask>=3.0.0
requests>=2.31.0
APScheduler>=3.10.4
PyGithub>=2.1.1
astroid>=3.0.0
networkx>=3.0
click>=8.0.0
python-dotenv>=1.0.0
